backend/requirements.txt
# ─── Core Framework ───────────────────────
fastapi==0.131.0
uvicorn==0.41.0
python-multipart==0.0.22

# ─── Data & Export ────────────────────────
pandas==3.0.1
openpyxl==3.1.5

# ─── Environment ─────────────────────────
python-dotenv==1.2.1

# ─── LLM Providers ───────────────────────
google-genai==1.64.0
openai==2.21.0

# ─── Resume Parsing ──────────────────────
PyMuPDF==1.27.1
PyPDF2==3.0.1
python-docx==1.2.0
scikit-learn==1.8.0

# ─── File Download ────────────────────────
gdown==5.2.1
requests==2.32.5
